Location, app logging, LinkedIn support
* LinkedIn ads work now, had to add some scraping stuff in a new
function
* logging which job opportunities I have here with employer, role,
location, url, notes. This is copy pasted into main spreadsheet tracking
where i've applied. For now that seems fine. It's on user to copy/paste
it at the right times, but with the "batch" workflow that's fine.
* Added "location" to JD_extractor. Turns out all the boards list a
location, so I should scrape that and save it! Not very useful, really,
but conforms to my existing spreadsheet model and mostly just tells me
about jobs in Oakland/Berkeley :)